Pull worker functionality out into a base class.
